requirements.txt
google-ai-generativelanguage
flask
gradio
pytest
python-dotenv